[MID-22] split SpecialistCreate into sub-components
