+++ src/app/admin/_components/specialist/SpecialistCreate/index.js
@@ -1,0 +1,3 @@
+import { SpecialistCreate } from './SpecialistCreate';
+
+export { SpecialistCreate };
