--- src/app/admin/_components/specialist/SpecialistList.js
@@ -1,5 +1,0 @@
-'use client';
-
-import React from 'react';
-
-export const SpecialistList = () => <p>For future tasks</p>;
