+++ src/app/admin/_components/specialist/index.js
@@ -1,4 +1,3 @@
 import { SpecialistCreate } from '@/app/admin/_components/specialist/SpecialistCreate';
-import { SpecialistList } from './SpecialistList';
 
-export { SpecialistList, SpecialistCreate };
+export { SpecialistCreate };
